Move models to a module.
